Feature: Prioritize characters of other players as opponents
When picking an opponent for a fight, we want to prioritize characters of other players, so a player's character has the least chances possible of fighting against a character of the same player.
Picking by rank difference is still the most important criteria, but ownership is more important than the fight count difference.